xu-ai-news-rag/rag-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
requests==2.32.3
